requirements.txt
numpy~=1.26
opencv_python~=4.10
pandas~=2.2
nicegui~=2.1
tensorflow~=2.17
joblib~=1.4
gudhi~=3.10
pillow~=10.4
pywebview~=5.2
torch~=2.4.1
torchvision~=0.19.1
opencv-python~=4.10.0.84
Flask~=3.0.3
Flask-RESTful~=0.3.10
aiohttp~=3.10.5
scikit-learn~=1.5.2
matplotlib~=3.9.2
pyserial~=3.5
python-sane~=2.9.1
